requirements.txt
discord.py[voice]>=2.7.1
PyNaCl
httpx
openai
pyyaml
apscheduler>=3.10.0
ollama
python-dotenv
typing_extensions
rich
requests
yfinance
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
azure-cognitiveservices-speech>=1.40.0
pydub>=0.25.1
pytest